// frontend/src/api/types.ts
export interface components {
    schemas: {
        LoginRequest: {
            email: string;
            password: string;
            role: string;
        };
        LoginResponse: {
            token: string;
        };
        RegisterRequest: {
            email: string;
            first_name: string;
            last_name: string;
            password: string;
        };
        /** @enum {string} */
        Role: "Admin" | "User";
        User: {
            email: string;
            first_name: string;
            /** Format: int32 */
            id: number;
            last_name: string;
            password: string;
            role: components["schemas"]["Role"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}